requirements.txt
# requirements.txt
fastapi
uvicorn
router
psycopg2-binary